register user

+++ src/components/Register.vue
@@ -1,6 +1,6 @@
 <template>
 
-<div class="container" id="login">
+<div class="container" id="register">
     <h2 class="green-text small">Register new user</h2>
     <div class="row">
       <div class="section col s2 m4"></div>
@@ -28,6 +28,9 @@
           </button>
           <RouterLink class="green-text register" to="/register">Login</RouterLink>
         </form>
+        <!-- ErrorMessage Printout -->
+        <div v-if="recievedMessage"><p class="text-green recievedMessage">{{recievedMessage}} </p></div>
+        <!-- <div v-if="errorMessage"><p class="text-green recievedMessage">{{errorMessage}} </p></div> -->
       </div>
       <div class="section col s2 m4"></div>
 
@@ -51,6 +54,12 @@ a.register {
   line-height: 36px;
 }
 
+.recievedMessage {
+  font-style: italic;
+  font-size: smaller;
+  margin-top: 1em;
+  line-height: 74px;
+}
 
 </style>
 
@@ -61,72 +70,99 @@ export default {
             name:"",
             email:"",
             password:"",
+            recievedMessage : null,
             errorMessage : null,
-            recievedMessage: null,
             user: null
         }
     },
     created : async function() {
         try {
             await this.addUser();
-            // this.token = response.data.token;
-            // console.log(this.token);
         }
         catch (error) {
             this.errorMessage = error;
+            window.localStorage.setItem('recievedMessage', error);
+            console.log(this.errorMessage);
         }
     },
     methods: {
         async addUser() {
-          
+
             // check first for content
             if(this.name.length > 2 ) {
                 let userBody = {
-                    name: this.name, 
-                    email: this.email, 
+                    name: this.name,
+                    email: this.email,
                     password: this.password
                 };
 
                 console.log(JSON.stringify(userBody));
-
+                try {
                 const response = await fetch("https://arcane-hamlet-64136.herokuapp.com/api/register",  {
                     method: "POST",
-                    headers: { 
+                    headers: {
                         "Accept" :  "application/json",
-                        "Content-type" : "application/json" 
+                        "Content-type" : "application/json"
                     },
                     body: JSON.stringify(userBody)
                 })
 
-
-                // .then(res => res.json())
-                // .then( data=> {
-                //     console.log(token);
-                // })
-                // .catch(err => console.log(error));
-
                 const data = await response.json();
-                console.log(data);
-                console.log(data.message);
-                console.log(data.error);
-                console.log(data.token);
-                console.log(data.user.name);
-                
+
+                console.log(response.status);
+
+                if(response.status === 200) {
+                  window.localStorage.setItem('token', data.token);
+                  window.localStorage.setItem('user', data.user.name);
+                  console.log(localStorage.getItem('user'));
+                  this.$emit("loggedin");
+                  // this.$router.push('/catalog');
+                  window.location.href = "/catalog";
+                }
+
+
+                window.localStorage.setItem('recievedMessage', data.error.email)
+                if(data.error.email == 'The email has already been taken.'){
+                    console.log(data.error.email);
+                    window.localStorage.setItem('recievedMessage', data.message);
+                    window.location.href = "/login";
+                };
+                this.recievedMessage = window.localStorage.getItem('recievedMessage');
+
+                if (response.status != 200) {
+                   throw (response.status);
+                }
+
+            }
+            catch (error) {
+                this.errorMessage = error;
+                console.log("catcherror: " + this.errorMessage);
+                // this.$router.push('/login');
+                // window.location.href = "/login";
+            }
+
                 // empty form
                 this.name = "";
                 this.email = "";
                 this.password = "";
-                
-                window.localStorage.setItem('token', data.token);
-                window.localStorage.setItem('recievedMessage', data.message);
-                window.localStorage.setItem('user', data.user.name);
+
+                // window.localStorage.setItem('token', data.token);
+                // window.localStorage.setItem('recievedMessage', data.message);
+                // window.localStorage.setItem('user', data.user.name);
                 // let token_ls = localStorage.getItem('token');
                 // console.log(token_ls);
 
-                this.$emit("addUser"); // reloads the parent page.
+                // this.$emit("addUser"); // reloads the parent page.
                 // window.location.href = "/catalog";
-                this.$router.push('/catalog');
-            }
+                //this.$router.push('/catalog');
+            } 
+            // else {
+            //     window.localStorage.setItem('recievedMessage', "Please fill in all fields.");
+            //     this.recievedMessage = "Please fill in all fields."
+            //     console.log(this.recievedMessage);
+            //     //this.$router.push('/register');
+            //     window.location.href = "/register";
+            // }
 
         }
     }
